Add BufferedHistory class for better OOP
