+++ src/main/java/seedu/address/history/History.java
@@ -33,6 +33,4 @@ public interface History<T> {
      * @return The current state.
      */
     T getCurrState();
-
-    T getCurrStateHasBuffer() throws HistoryException;
 }
